mirror trxcllnt/js-data-refactor#8dc5d2cbf12de0e2adbe2b85e73ec7d5e80e9827

+++ type.d.ts
@@ -187,7 +187,7 @@ declare type FType = {
     };
     [Type.Float64]: {
         precision: Precision.DOUBLE;
-        TArray: Float32Array;
+        TArray: Float64Array;
         TValue: number;
     };
 };
